+++ ChangeLog.md
@@ -4,7 +4,8 @@ This changelog summarizes changes and fixes which are a part of each revision.  
 where provided, and the git commit history.
 
 * 2017/06/?? - 4.8 - PENDING
-    *
+    * Changes to Create, View, and Edit concept to support VHAT usage. Includes label changes, dynamic extended type description options in create and edit, hiding or showing description type,
+       extended description type, and dialects depending on the terminology type, and other bug fixes. (Jazz: 529741)
     
 * 2017/06/08 - 4.7
     * Added Generate VUID dialog (Jazz #501723)